Create SingleValue<> struct for event that handles primitive paramiters like int
This is because in the inspector when it's an int it force us to indicate the value of the parameter (which we don't want because we want to get the value from the event).

So I created the wrapper struct to hold the value.

I don't know if there's a better solution.